+++ packages/esbuild-pilet-plugin/src/index.ts
@@ -2,7 +2,7 @@ import type { SharedDependency } from 'piral-cli';
 import { Plugin } from 'esbuild';
 import { transformFileAsync } from '@babel/core';
 import { promises } from 'fs';
-import { isAbsolute, join, resolve, basename } from 'path';
+import { resolve, basename } from 'path';
 
 export interface PiletPluginOptions {
   name: string;
@@ -13,51 +13,7 @@ export interface PiletPluginOptions {
 export const piletPlugin = (options: PiletPluginOptions): Plugin => ({
   name: 'pilet-plugin',
   setup(build) {
-    const loaders = build.initialOptions.loader || {};
-    const extensions = [];
-
-    for (const ext of Object.keys(loaders)) {
-      const loader = loaders[ext];
-
-      if (loader === 'file') {
-        extensions.push(ext);
-        delete loaders[ext];
-      }
-    }
-
     build.initialOptions.metafile = true;
-    const filter = new RegExp(`(${extensions.map((ext) => `\\${ext}`).join('|')})$`);
-
-    build.onResolve({ filter }, (args) => {
-      if (args.namespace === 'ref-stub') {
-        return {
-          path: args.path,
-          namespace: 'ref-binary',
-        };
-      } else if (args.resolveDir !== '') {
-        return {
-          path: isAbsolute(args.path) ? args.path : join(args.resolveDir, args.path),
-          // for CSS we'll just use the path; no intermediate module needed
-          namespace: args.kind === 'url-token' ? 'ref-binary' : 'ref-stub',
-        };
-      } else {
-        return; // Ignore unresolvable paths
-      }
-    });
-
-    build.onLoad({ filter: /.*/, namespace: 'ref-stub' }, async (args) => ({
-      resolveDir: resolve(__dirname),
-      contents: [
-        `import path from ${JSON.stringify(args.path)}`,
-        `import { __bundleUrl__ } from ${JSON.stringify('../../set-path.js')}`,
-        `export default __bundleUrl__ + path;`,
-      ].join('\n'),
-    }));
-
-    build.onLoad({ filter: /.*/, namespace: 'ref-binary' }, async (args) => ({
-      contents: await promises.readFile(args.path),
-      loader: 'file',
-    }));
 
     build.onEnd(async (result) => {
       const root = process.cwd();