+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.15.1 (tbd)
 
 - Updated Lerna to v6
+- Added `esbuild-auto-path-plugin` package
 
 ## 0.15.0 (November 17, 2022)